arsketch_backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
Pillow==9.2.0